backend/requirements.txt
# CodeIntegrity Backend Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
motor>=3.3.0
pymongo>=4.6.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1
python-multipart>=0.0.6

# Redis
redis>=5.0.0

# Analysis Engine
pygments>=2.17.0
pycparser>=2.21
javalang>=0.13.0
tree-sitter>=0.21.0
tree-sitter-java>=0.21.0
tree-sitter-cpp>=0.21.0
mmh3>=4.0.0

# Email
requests>=2.31.0

# Encryption
cryptography>=42.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
tzdata
